Url validator fix
